Column.validate is not working with cellEditable #2621
